+++ packages/terra-grid/tests/wdio/grid-spec.js
@@ -1,5 +1,3 @@
-/* global browser, Terra, before */
-
 describe('Grid', () => {
   before(() => browser.setViewportSize(Terra.viewports('medium')[0]));